+++ utils.py
@@ -1,0 +1,7 @@
+from enum import StrEnum
+
+class KnowledgeBaseCollection(StrEnum):
+    """
+    Enum for the different knowledge base collections.
+    """
+    GENERIC = "generic"